fix(transcription): break select<->cli import cycle that crashed app boot
The engine-selection extraction left whisper-cli/parakeet-cli importing catalogEngine/
modelsByEngine from select.ts, while select reads the service singletons at module scope
(const ALL). On the bundled main-process load order that cycle threw 'Cannot access
transcriptionService before initialization' - the app would not start, yet tsc + unit tests +
build all passed (the TDZ only shows in the bundled load order). Move the pure classifiers to
a leaf module classify.ts that the CLIs import; select re-exports them for compatibility. Adds
a structural guard test so the cycle cannot silently return.
